Files for Voron1 issue
Problem and files shared here:
https://github.com/VoronDesign/Voron-1/issues/63



IEC Doesn't Match BOM + Workaround Adam Tech IEC-GS-1-100


IEC power inlet no longer matches the BOM. 

Additionally, I couldn't find the 300 mm power skirt in the CAD model.

There's a few shortcomings in the CAD model. Primarily, the version included seems to be a zero-clearance fit around the intended part and is not removable.

Workaround:
On Discord, I and a few other users mentioned purchasing an Adam Tech IEC-GS-1-100, currently the part for the V0.

I'm attaching the STLs and CAD model for an updated version. The 300 mm version fits the Adam Tech IEC plug with the recommended cut-out dimensions. It engages with a loose fit and provides rear access to remove the retaining clips if needed.

Known deficiencies:
The CAD models I've attached have a fairly thin feature at the top. Ideally the plug should be slid down a mm or so. That said, the thin wall printed just fine for me.

I have provided the 250 mm version but have not tested it.